move base classes and add a todo
