+++ testomania/src/main/java/com/earth/testomania/quiz_categories/AnimalsDestination.kt
@@ -5,6 +5,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.earth.testomania.apis.quiz.opentdb.domain.models.OpenTDBApiCategory
 import com.earth.testomania.apis.quiz.opentdb.domain.repository.OpenTdbRepo
 import com.earth.testomania.quiz_categories.usecase.OpenTDBApiBaseUrlUseCase
+import com.earth.testomania.quiz_categories.viewmodel.DestinationViewModel
 import com.earth.testomania.quiz_screen.MainQuizScreen
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
