+++ testomania/src/main/java/com/earth/testomania/quiz_categories/InformationalTechnologiesDestination.kt
@@ -5,6 +5,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.earth.testomania.apis.quiz.quizapi.domain.model.QuizApiCategory
 import com.earth.testomania.apis.quiz.quizapi.domain.repository.QuizRepository
 import com.earth.testomania.quiz_categories.usecase.QuizApiBaseUrlUseCase
+import com.earth.testomania.quiz_categories.viewmodel.DestinationViewModel
 import com.earth.testomania.quiz_screen.MainQuizScreen
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
