+++ testomania/src/main/java/com/earth/testomania/quiz_screen/ui_components/CategoryIllustration.kt
